+++ packages/react-server/src/server/test/server.test.tsx
@@ -1,7 +1,8 @@
 import React from 'react';
 import {useTitle} from '@shopify/react-html';
-import {useRequestHeader, Header} from '@shopify/react-network';
+import {useCookie, CookieUniversalProvider} from '@shopify/react-cookie';
 import {createServer} from '../server';
+
 import {mockMiddleware, TestRack} from '../../test/utilities';
 
 const rack = new TestRack();
@@ -56,19 +57,38 @@ describe('createServer()', () => {
     );
   });
 
-  it('supports getting headers', async () => {
+  it('supports getting cookies', async () => {
+    const cookie = 'foo';
+    const value = 'bar';
+
     function MockApp() {
-      const header = useRequestHeader(Header.UserAgent);
-      return <div>{header}</div>;
+      const [cookieValue] = useCookie(cookie);
+
+      return <>{cookieValue}</>;
     }
 
-    const wrapper = await rack.mount(({ip, port}) =>
-      createServer({port, ip, render: () => <MockApp />}),
+    const wrapper = await rack.mount(
+      ({ip, port}) =>
+        createServer({
+          port,
+          ip,
+          render: () => (
+            <CookieUniversalProvider server>
+              <MockApp />
+            </CookieUniversalProvider>
+          ),
+        }),
+      {
+        headers: {
+          cookie: `${cookie}=${value}`,
+        },
+      },
     );
+
     const response = await wrapper.request();
 
     expect(await response.text()).toStrictEqual(
-      expect.stringContaining(`node-fetch`),
+      expect.stringContaining(`<div id="app">${value}</div>`),
     );
   });
 });